Reverted changes with jest mocking.
